Re-organize spec, make tests more readable.
A major change from script-like functions for writing configuration to
a builder approach.
There are builders for rules, schedulers and interpreters.
Exec and mock expectations are also made through the builders.

test.common.mock has a test now and works through multiple expectation
groups in parallel.
This is necessary for builders to be able to accumulate their
expectations.

main.py uses a pymoduleloader instead of a pymoduleschedulerloader
-> difference: spec has to account for init calls to schedulers