Move locking to core and expand it.
Make the SyncRule acquire an fcntl lock in a special lock dir before
syncing. This will duplicate the effort of schedulers that lock anyway
(like anacron), but it is necessary because execute-rule is now
supposed to be usable in a stand-alone fashion -- without a scheduler.

To support delayed as well as stand-alone executions, make the sibt sync
action re-validate the rule it is given.

Because the restore/versions-of/list-files actions work with sys rules,
which the root user might be syncing at the moment, locking is out of
the question here. So schedulers are supposed to predict the next
execution in their nextExecutionTime function. Whenever this is too
close, an unstable phase detector tells a SyncRule that it is unstable.
The SyncRule will then abort a restoration.

Also make two major changes in the acceptance tests:
  * Transfer some specs to schedulerusagespec.py to clean up
    sibtspec.py, contains specs that involve the specific
    implementation of scheduler functions. The quick_acceptance_test
    task in pavement.py runs all acceptance tests except the
    processspec.
  * Treat sysPaths as user paths if setRootUserId is used. Default for
    a test is setNormalUserId (was previously setRootUserId). That way,
    a single test can simulate both users with a single sys config.

Changes:
  * Add test folder to python path in pavement.py. Is already done by
    py.test, but that does not update the PYTHONPATH env var.
  * Construct SyncRules with the correct log. Now they can provide
    attributes like nextExecution and latestExecution.
  * See above. versions-of/list-files warns if a rule whose ports are
    affected is unstable.
  * An execution is considered close if it is going to happen or
    happened in less than 1 hour. Chosen because anacron is usually run
    hourly.
  * restore --force will not abort if rule is unstable.
  * Implement nextExecutionTime in anacron.
  * If --verbose is passed, execute-rule will log to stderr.
  * Change cliparser to not choose the default action (list) if only
    rule names are given.

Refactorings:
  * Rename the sync-uncontrolled action to sync. The uncontrolled part
    mostly referred to the absence of locking, which is gone now.