Implement mount point assertions and simple sched.
Add a new validator MountPointAssertionsTrueValidator and a rule option
MustBeMountPoint. The latter takes a list of loc numbers that the
former iterates over and tests with os.path.ismount. The rationale is as
follows. A loc might be accidentally unmounted, because a script line was
forgotten, or because some of the filesystem's blocks are damaged. If
the assigned loc path is for example /mnt/disk/foo, and /mnt/disk is
the mount point, then the loc path does not exist and the error is
caught. If, however, the path was /mnt/disk, then the synchronizer
would be called with an empty directory, which could result in data
loss, and, worse, the error would go unnoticed. MustBeMountPoint can
prevent this.

MustBeMountPoint is tested by mounting a test FUSE. The test is skipped
if fuse cannot be found. Note that, currently, fuse-python is python 2
only, so another interpreter has to be called.

The simple sched is not as simple as possible, but reasonably
bare-bones while being usable. Its schedule immediately calls
execute-rule for all due schedulings in parallel. It has no options except
Interval, which can be any timedelta. The next execution time is the
last time without time of day plus interval. Time of day is removed
to prevent the start time of day from wandering on a continuously
running machine.

Changes:
  * anacron: Add OutputTabFile option. Anacron writes the tab to this
    path and does not execute if given.
  * The scheduler args include a clock because one is generally
    necessary when implementing nextExecutionTime and to make stuff
    testable. Note that by mutating the clock and calling
    nextExecutionTime more than one execution could be predicted.
  * Add attribute Scheduling.lastExecutionTime. Necessary in simple
    scheduler for calculating next time and to decide whether to execute.
  * Remove restriction that prevented FatalSignalExceptions to be
    thrown twice. Subsequent SIGTERMs are still ignored because sibt
    signals its process group in this case. Use case: simple waits
    for its children, then again after being signaled, but not a third
    time.
  * Lock in SyncRule.execute, not SyncRule.sync. The simple sched does
    not know if a rule is executing when calling execute-rule. It
    should not generate failed execution loggings each time. Does not
    affect anacron. Note that sync is again uncontrolled, actually.
  * Make the FilesDBExecutionsLog return finished and failed executions
    if the log file is not being written to any more (power outage,
    SIGKILL, ...). It does it by lockf (3)ing and testing the lock.
  * LoggingScheduler: Treat Syslog as on by default.

Refactorings:
  * Remove SchedulingBuilder because it is not pythonic.
  * Move code common to simple and anacron tests to leafschedulertest.