Implement tar syncer with SSHFS automounting.
The old tar synchronizer stub written in python is moved to pytar and
should be removed. The new tar syncer uses GNU tar's incremental backup
feature. It makes a full backup from time to time and incremental ones
in between, always keeping one new and one old full backup around.
Keeping two is a compromise between data security and memory
consumption. Overwriting an existing backup is a bad idea because the
process could be interrupted, leaving the user with no data at all. Care
was taken to make the mv(1)s when syncing signal and power outage
proof, but it is hard to test.

An increment is accompanied by a snapshot file. “It contains the status
of the file system at the time of the dump”. Version 2 is current and
has been since tar 1.16. versionsOf and listFiles work by parsing
these snapshots because calling --list on incremental archives outputs
only the changed files and would be tedious to parse anyway because of tar
quoting every file name.

Note that the parser is written in python as bash is simply too slow,
but even though, it is not as fast as it could be. Parsing a 30M
snapshot file from a system backup takes 4s while grepping over the
corresponding 90M plain text file list takes 0.3s. But grepping is not an
option, as I think it cannot do everything listFiles needs to do.
Note also that if the parser could use the file.readline() method,
it would take only 1s in the aforementioned experiment. The snapshots
use null bytes as delimiters, though, not newlines.

GNU tar has the option to use a remote --file. It would then connect via
rsh (ssh) to the rmt(8) daemon running on the target host. This proved
to be slower than just mounting via sshfs and running tar on local
files in an experiment of mine, however. So any synchronizer not
supporting ssh will get the ability when wrapped with the
SSHFSAutomountingSynchronizer, which mounts and unmounts around calls to
listFiles, versionsOf, restore, and sync. The whole setup is tested
with the sshserver, now extended with a logging null implementation of the
paramiko SFTP subsystem.

Changes:
  * In assertutil: Print iterable as assertion message because py.test
    sometimes cannot re-write the assertions properly.
  * Remove support for parsing W3C date times in versionsOf output of
    syncers. If present, W3C strings can easily be translated to
    timestamps with the date(1) program.
  * Add a milliseconds part to timestamps passed to and from syncers, i.e.,
    123,0-999 now instead of 123. Note that the milliseconds will not
    be displayed in sibt's front end. The main purpose of this is to
    make syncer tests faster because there is no need to wait for one
    second to get another timestamp anymore now. Change tests
    accordingly.
  * Make scheduler and synchronizer loading lazy. Both are packed into
    a ConfigurableList which has a getByName method that loads, caches,
    and returns the result. Thus, scheds and syncers not used by rules
    will not be called. Important in tests that mock test syncer calls.
  * Write BashFuncFixture and use it in runner_test and syncer tests.
    It calls functions from a sourced bash script so the result can be
    tested against.
  * rsync test: Test ExcludedDirs reanchoring function separately.
  * sibtspec: Fix \0 printing in one test that went unnoticed:
    echo -n $'foo\0' does not work, echo -n foo; echo -n -e '\0' does.
  * Adapt rdiff-backup test and fix bash read calls. Should always be
    called as read -r.
  * rdiff-backup: Ignore test that syncs to symlink. Not possible for
    now.
  * Parallelize versions-of: Necessary now with SSHFS automounting and
    tar taking its time. As many threads as rules are spawned and a
    parallel map is computed. Non-blocking, selector-based, or
    asyncio-based process runner implementations would be possible,
    but harder to do.

Refactorings:
  * Do not link to /home for testing purposes in SynchronizerTest
    because it might be written to.
  * rsync and rsync test: Share ExcludedDirs code and test with tar
    in ExcludingSynchronizerTest.
  * Split SynchronizerTest into Listing test and Restoring test so
    newly written syncers can inherit from one after the other until
    everything is implemented.
  * Move options to the front of syncer listFiles, versionsOf, and
    restore function parameter lists for easier currying. Used in
    SSHFSAutomountingSyncer: def versionsOf(self, options, *args)
    Adapt tests (sibtspec, runners_test, synchronizertest).
  * Prefix private functions in bash synchronizers with -.