Integrate duplicity and fix some bugs.
Write a synchronizer bash script to interface with duplicity and test it
with local files as well as a WebDAV server. SSH URLs are supported as
well, and they default to using the sftp duplicity backend since the
scp backend is not well-maintained, for example not being able to handle
single quotes in file names. However, the -oUserKnownHostsFile does not
get parsed by the backend which is why automated testing of scp is currently
not possible. Testing sftp is also not possible because the test server
does not support it. Both can be fixed. In the meantime, sftp has been
tested with a local OpenSSH server. Also, test encryption.

Bug fixes:
  * rsync and rdiff-backup: If RemoteShellCommand is not set, construct one
    if a socket port is set. Unsusal scenario.
  * Update SynchronizerOptions.options when an item is set. Otherwise,
    calling withNewLocs would reset the options to old values.
  * Create rdiff-backup-data/restore.log after sync() so a read-only
    restorer does not need to write into the folder. Also, make it
    world-writable.
  * Prevent crashes from trying to decode arbitrary bytes with the system
    encoding. Possible solution: Simply do not decode. However, the Python
    way is to use errors="surrogateescape" everywhere an OS buffer is input
    or needs to be output. Do this here:
      - in coprocessrunner
      - inifilereader
      - in fileobjoutput
      - already done implicitly in sys.argv
    Change from text to binary: stdout and stderr ->
      rename unbufferedtextfile, use pytest capfdbinary.

Changes:
  * Do not check locNumber=2 list-files in synchronizers and tests,
    UnidirectionalSyncerTest now tests that versions-of returns nothing if
    locNumber=2. Since list-files is only called with output from
    versions-of, this is sufficient.
  * Instead of rounding up to 60m ago, show 1h0m ago.
  * Sort versions-of output after it is gathered from the syncers.

Refactorings:
  * Rename SynchronizerTestFixture -> RunnableFileSynchronizerTestFixture and
    so SynchronizerTestFixture can be shared with
    SSHFSAutomountingSyncer test.
  * sshfsautomountingsynchronizer: Special value None -> _FileMountPoint